requirements.txt
fastapi
uvicorn
pandas
openpyxl
ydata-profiling
langgraph
openai
matplotlib
altair
seaborn
reportlab
python-multipart
python-dotenv
